Sorting and removing files works. need to update styles
